Impl next_power_of_tw
